requirements.txt
asgiref==3.8.1
boto3==1.38.44
botocore==1.38.44
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
jmespath==1.0.1
pillow==11.2.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.4
rest-framework-simplejwt==0.0.2
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
